move socket tests to apptest_socket to debug failure, fix
Also add appropriate tests to apptest_exctable_dis to catch
a regression earlier
